Move Agent, Landmark and Victim to their respective files
